Add files via upload
This is a Camera-App with many features inspired with Smart phones camera. Those features are as follow :- 
1.) Image Capturing
2.) Audio & Video Recording
3.) Zoom-in & Zoom-out
4.) Filters.

Web_Camera-App is created without using Image-processing but the incredients are :- HTML, CSS, JavaScript. The tools that used in javascript is canvas, mediaStream etc ; In CSS animation also are implemented ; The Video tag has the prior role in the framing of HTML. Users can also download the recorded video and captured image with filters easily by one click only. 

Soon the concept of indexedDB will gone to be introduced in this app for making the DataBase as gallery in the browser so that the users can capture images or record videos as much as they can and then they can download those images or videos they want.